refactor(dataobj): detach StreamsDecoder from Decoder
This change allows a StreamsDecoder to be created outside of the context
of a Decoder, instead using a SectionReader. This will assist in moving
the implementation of StreamsDecoder to the streams package.
